7. Winning is only half of it. Having fun is the other half

+++ pointers_arrays_strings/7-puts_half.c
@@ -3,7 +3,7 @@
 void puts_half(char *str)
 {
 	int i;
-	size_t strlen(const char *str);
+	int strlen(const char *str);
 
 	for (i = strlen(str) / 2; i < strlen(str); i++)
 		_putchar(str[i]);
